Add tooltip explaining data replacement
